Support listening to changes
The option listen (any value for now) will install a transaction
listener that will re-run the query when new transactions come in.

The query must be deferred for listen to be used. The same process
will receive the new results. When the listening process stops for any
reason, the listener will be removed.